BillSplit components splitting into two-Group part working
